+++ android/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `app` module which will be the launch configuration for the OSSP Project.
 - `core` feature module which contains all core services
 - Added a Readme and this Changelog to track major changes
+- Added database and injection module to core.
+- Added User and Auth tables aligning with [Architecture]
 
 [Unreleased]: https://github.com/CCorrado/ssw690ossmgmt/tree/master/android
 [0.0.1]: https://github.com/CCorrado/ssw690ossmgmt/tree/master/android
+[Architecture]: https://github.com/CCorrado/ssw690ossmgmt/blob/master/ossp_arch.png